requirements.txt
alembic==1.4.3
azure-identity==1.7.0
azure-keyvault-keys==4.5.1
azure-keyvault-secrets==4.4.0
azureml-core==1.44.0
azureml-mlflow==1.44.0
class_registry==2.1.2
cron-descriptor==1.2.31
cron-validator==1.0.6
numpy==1.22.4
# d3rlpy>=1.1.1
dask[complete]==2022.8.0
gurobipy==10.0.0
hydra-core==1.2.0
lightgbm==3.3.2
matplotlib==3.5.2
networkx==2.8.5
numba==0.56.0
ortools==9.3.10497
psycopg2-binary==2.9.3
pydantic[dotenv]==1.9.1
pytest==7.1.3
pytorch-lightning==1.7.1
ray[air]
ray[all]==2.2.0
rsome==1.1.2
sb3-contrib
scikit-learn==1.2.0
seaborn==0.11.2
SQLAlchemy==1.3.19
sqlalchemy-utils==0.36.8
stable-baselines3
typer-cli==0.0.12
wandb==0.13.1
windpowerlib==0.2.1
